requirements.txt
aiohttp==3.9.5
aiofiles==23.2.1
asyncio==3.4.3
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.20.0
playwright==1.44.0
redis==5.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
pandas==2.2.2
numpy==1.26.4
schedule==1.2.1
python-dotenv==1.0.1
loguru==0.7.2
influxdb-client==1.41.0
grafana-api==1.0.3
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4